fix: types for errors and warnings
