+++ lib/Cache.js
@@ -11,8 +11,6 @@ const {
 	makeWebpackErrorCallback
 } = require("./errors/HookWebpackError");
 
-/** @typedef {import("./errors/WebpackError")} WebpackError */
-
 /**
  * Cache validation token whose string representation identifies the build
  * inputs associated with a cached value.
@@ -21,11 +19,10 @@ const {
  */
 
 /**
- * Completion callback used by cache operations that either fail with a
- * `WebpackError` or resolve with a typed result.
+ * Completion callback used by cache operations that either fail with a `Error` or resolve with a typed result.
  * @template T
  * @callback CallbackCache
- * @param {WebpackError | null} err
+ * @param {Error | null} err
  * @param {T=} result
  * @returns {void}
  */
