+++ lib/Compilation.js
@@ -3209,7 +3209,7 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 				let hasProblems = this.reportDependencyErrorsAndWarnings(module, [
 					module
 				]);
-				const errors = module.getErrors();
+				const errors = /** @type {WebpackError[]} */ (module.getErrors());
 				if (errors !== undefined) {
 					for (const error of errors) {
 						if (!error.module) {
@@ -3219,7 +3219,7 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 						hasProblems = true;
 					}
 				}
-				const warnings = module.getWarnings();
+				const warnings = /** @type {WebpackError[]} */ (module.getWarnings());
 				if (warnings !== undefined) {
 					for (const warning of warnings) {
 						if (!warning.module) {
