+++ lib/Module.js
@@ -303,9 +303,9 @@ class Module extends DependenciesBlock {
 		/** @type {boolean} */
 		this.hot = false;
 		// Info from Build
-		/** @type {WebpackError[] | undefined} */
+		/** @type {Error[] | undefined} */
 		this._warnings = undefined;
-		/** @type {WebpackError[] | undefined} */
+		/** @type {Error[] | undefined} */
 		this._errors = undefined;
 		/** @type {BuildMeta | undefined} */
 		this.buildMeta = undefined;
@@ -782,7 +782,7 @@ class Module extends DependenciesBlock {
 
 	/**
 	 * Adds the provided warning to the module.
-	 * @param {WebpackError} warning the warning
+	 * @param {Error} warning the warning
 	 * @returns {void}
 	 */
 	addWarning(warning) {
@@ -794,7 +794,7 @@ class Module extends DependenciesBlock {
 
 	/**
 	 * Returns list of warnings if any.
-	 * @returns {Iterable<WebpackError> | undefined} list of warnings if any
+	 * @returns {Error[] | undefined} list of warnings if any
 	 */
 	getWarnings() {
 		return this._warnings;
@@ -810,7 +810,7 @@ class Module extends DependenciesBlock {
 
 	/**
 	 * Adds the provided error to the module.
-	 * @param {WebpackError} error the error
+	 * @param {Error} error the error
 	 * @returns {void}
 	 */
 	addError(error) {
@@ -822,7 +822,7 @@ class Module extends DependenciesBlock {
 
 	/**
 	 * Returns list of errors if any.
-	 * @returns {Iterable<WebpackError> | undefined} list of errors if any
+	 * @returns {Error[] | undefined} list of errors if any
 	 */
 	getErrors() {
 		return this._errors;
@@ -1393,13 +1393,13 @@ Object.defineProperty(Module.prototype, "errors", {
 	/**
 	 * Returns errors.
 	 * @deprecated
-	 * @returns {WebpackError[]} errors
+	 * @returns {Error[]} errors
 	 */
 	get: util.deprecate(
 		/**
 		 * Returns errors.
 		 * @this {Module}
-		 * @returns {WebpackError[]} errors
+		 * @returns {Error[]} errors
 		 */
 		function errors() {
 			if (this._errors === undefined) {
@@ -1417,13 +1417,13 @@ Object.defineProperty(Module.prototype, "warnings", {
 	/**
 	 * Returns warnings.
 	 * @deprecated
-	 * @returns {WebpackError[]} warnings
+	 * @returns {Error[]} warnings
 	 */
 	get: util.deprecate(
 		/**
 		 * Returns warnings.
 		 * @this {Module}
-		 * @returns {WebpackError[]} warnings
+		 * @returns {Error[]} warnings
 		 */
 		function warnings() {
 			if (this._warnings === undefined) {
