+++ lib/NormalModule.js
@@ -30,7 +30,7 @@ const ModuleError = require("./errors/ModuleError");
 const ModuleParseError = require("./errors/ModuleParseError");
 const ModuleWarning = require("./errors/ModuleWarning");
 const UnhandledSchemeError = require("./errors/UnhandledSchemeError");
-const WebpackError = require("./errors/WebpackError");
+const NonErrorEmittedError = require("./errors/NonErrorEmittedError");
 const LazySet = require("./util/LazySet");
 const { isSubset } = require("./util/SetHelpers");
 const { getScheme } = require("./util/URLAbsoluteSpecifier");
@@ -207,24 +207,6 @@ const asBuffer = (input) => {
 	return input;
 };
 
-class NonErrorEmittedError extends WebpackError {
-	/**
-	 * @param {EXPECTED_ANY} error value which is not an instance of Error
-	 */
-	constructor(error) {
-		super();
-
-		this.name = "NonErrorEmittedError";
-		this.message = `(Emitted value instead of an instance of Error) ${error}`;
-	}
-}
-
-makeSerializable(
-	NonErrorEmittedError,
-	"webpack/lib/NormalModule",
-	"NonErrorEmittedError"
-);
-
 /** @typedef {[string | Buffer, string | RawSourceMap | undefined, PreparsedAst | undefined]}  Result */
 
 /** @typedef {LoaderContext<EXPECTED_ANY>} AnyLoaderContext */
@@ -394,7 +376,7 @@ class NormalModule extends Module {
 		this.extractSourceMap = extractSourceMap;
 
 		// Info from Build
-		/** @type {WebpackError | null} */
+		/** @type {Error | null} */
 		this.error = null;
 		/**
 		 * @private
@@ -1176,7 +1158,7 @@ class NormalModule extends Module {
 	}
 
 	/**
-	 * @param {WebpackError} error the error
+	 * @param {Error} error the error
 	 * @returns {void}
 	 */
 	markModuleAsErrored(error) {
@@ -1328,7 +1310,7 @@ class NormalModule extends Module {
 				try {
 					hooks.beforeSnapshot.call(this);
 				} catch (err) {
-					this.markModuleAsErrored(/** @type {WebpackError} */ (err));
+					this.markModuleAsErrored(/** @type {Error} */ (err));
 					return callback();
 				}
 
@@ -1420,7 +1402,7 @@ class NormalModule extends Module {
 			try {
 				hooks.beforeParse.call(this);
 			} catch (err) {
-				this.markModuleAsErrored(/** @type {WebpackError} */ (err));
+				this.markModuleAsErrored(/** @type {Error} */ (err));
 				this._initBuildHash(compilation);
 				return callback();
 			}
