+++ lib/Watching.js
@@ -11,7 +11,6 @@ const Stats = require("./Stats");
 /** @typedef {import("./Compilation")} Compilation */
 /** @typedef {import("./Compiler")} Compiler */
 /** @typedef {import("./Compiler").ErrorCallback} ErrorCallback */
-/** @typedef {import("./errors/WebpackError")} WebpackError */
 /** @typedef {import("./logging/Logger").Logger} Logger */
 /** @typedef {import("./util/fs").TimeInfoEntries} TimeInfoEntries */
 /** @typedef {import("./util/fs").WatchFileSystem} WatchFileSystem */
@@ -480,7 +479,7 @@ class Watching {
 		}
 		/**
 		 * Processes the provided err.
-		 * @param {WebpackError | null} err error if any
+		 * @param {Error | null} err error if any
 		 * @param {Compilation=} compilation compilation if any
 		 */
 		const finalCallback = (err, compilation) => {
@@ -495,7 +494,7 @@ class Watching {
 			this.compiler.fsStartTime = undefined;
 			/**
 			 * Processes the provided err.
-			 * @param {WebpackError | null} err error if any
+			 * @param {Error | null} err error if any
 			 */
 			const shutdown = (err) => {
 				this.compiler.hooks.watchClose.call();
