+++ lib/errors/HookWebpackError.js
@@ -85,7 +85,7 @@ module.exports.makeWebpackError = makeWebpackError;
 /**
  * Creates webpack error callback.
  * @template T
- * @param {(err: WebpackError | null, result?: T) => void} callback webpack error callback
+ * @param {(err: Error | null, result?: T) => void} callback webpack error callback
  * @param {string} hook name of hook
  * @returns {Callback<T>} generic callback
  */
